Refactor component graphs into a combinatorial core
Introduce a new purely combinatorial version of
ComponentGraph, modeling component graphs as
multigraphs with vertex genera and derived one-tails
(elliptic tails and pigtails). Add canonical
signatures for isomorphism testing and classification,
and provide human-readable output via as_text().

The construction of component graphs from concrete
plane curves is moved to a separate module
(component_graph_of_plane_curve.py), which is added
in a preliminary, not-yet-functional state and
will be completed later.
